Remove deprecated Atlassian installation routes and tests, consolidating functionality into connectors-atlassian routes. Update related mocks in tests for improved organization and clarity.
