+++ apps/backend/src/routes/v1/connectors-atlassian.test.ts
@@ -14,13 +14,26 @@ const getForgeInstallationByOrgIdMock = vi.hoisted(() => vi.fn())
 const listConfluenceSelectionsByOrgIdMock = vi.hoisted(() => vi.fn())
 const getAtlassianUserAccessTokenMock = vi.hoisted(() => vi.fn())
 const replaceConfluenceSelectionsMock = vi.hoisted(() => vi.fn())
+const getPendingForgeInstallationForUserInOtherOrgMock = vi.hoisted(() =>
+  vi.fn(),
+)
+const upsertPendingForgeInstallationMock = vi.hoisted(() => vi.fn())
 
-vi.mock("../../models/atlassian-connector.js", () => ({
-  getForgeInstallationByOrgId: getForgeInstallationByOrgIdMock,
-  listConfluenceSelectionsByOrgId: listConfluenceSelectionsByOrgIdMock,
-  getAtlassianUserAccessToken: getAtlassianUserAccessTokenMock,
-  replaceConfluenceSelections: replaceConfluenceSelectionsMock,
-}))
+vi.mock("../../models/atlassian-connector.js", async (importOriginal) => {
+  const actual = await importOriginal<
+    typeof import("../../models/atlassian-connector.js")
+  >()
+  return {
+    ...actual,
+    getForgeInstallationByOrgId: getForgeInstallationByOrgIdMock,
+    listConfluenceSelectionsByOrgId: listConfluenceSelectionsByOrgIdMock,
+    getAtlassianUserAccessToken: getAtlassianUserAccessTokenMock,
+    replaceConfluenceSelections: replaceConfluenceSelectionsMock,
+    getPendingForgeInstallationForUserInOtherOrg:
+      getPendingForgeInstallationForUserInOtherOrgMock,
+    upsertPendingForgeInstallation: upsertPendingForgeInstallationMock,
+  }
+})
 
 import { requireOrgAdminOrOwner } from "../../auth/withAuth.js"
 import { atlassianConnectorRoutes } from "./connectors-atlassian.js"
@@ -48,6 +61,18 @@ describe("Atlassian connector routes", () => {
     listConfluenceSelectionsByOrgIdMock.mockResolvedValue([])
     getAtlassianUserAccessTokenMock.mockResolvedValue(undefined)
     replaceConfluenceSelectionsMock.mockResolvedValue([])
+    getPendingForgeInstallationForUserInOtherOrgMock.mockResolvedValue(
+      undefined,
+    )
+    upsertPendingForgeInstallationMock.mockResolvedValue({
+      id: "fgi_default",
+      orgId: "org_1",
+      cloudId: null,
+      status: "pending",
+      installedByUserId: "user_1",
+      createdAt: new Date("2026-03-01T00:00:00.000Z"),
+      updatedAt: new Date("2026-03-01T00:00:00.000Z"),
+    })
   })
 
   it("GET /status returns connector state", async () => {
@@ -64,7 +89,7 @@ describe("Atlassian connector routes", () => {
     expect(await res.json()).toMatchObject({
       isLinked: true,
       isInstalled: true,
-        installationStatus: "installed",
+      installationStatus: "installed",
       selectedPageCount: 1,
     })
   })
@@ -98,5 +123,68 @@ describe("Atlassian connector routes", () => {
     })
     expect(res.status).toBe(200)
     expect(await res.json()).toEqual({ savedCount: 2 })
+  })
+
+  describe("POST /connectors/atlassian/installation", () => {
+    beforeEach(() => {
+      getAtlassianUserAccessTokenMock.mockResolvedValue("atl_token")
+      getPendingForgeInstallationForUserInOtherOrgMock.mockResolvedValue(
+        undefined,
+      )
+      upsertPendingForgeInstallationMock.mockResolvedValue({
+        id: "fgi_1",
+        orgId: "org_1",
+        cloudId: null,
+        status: "pending",
+        installedByUserId: "user_1",
+        createdAt: new Date("2026-03-01T00:00:00.000Z"),
+        updatedAt: new Date("2026-03-01T00:00:00.000Z"),
+      })
+    })
+
+    it("returns 409 when Atlassian account is not linked", async () => {
+      getAtlassianUserAccessTokenMock.mockResolvedValueOnce(undefined)
+      const app = createApp()
+      const res = await app.request("/connectors/atlassian/installation", {
+        method: "POST",
+      })
+
+      expect(res.status).toBe(409)
+      expect(await res.json()).toEqual({
+        error: "Atlassian account not linked",
+        code: "atlassian_not_linked",
+      })
+    })
+
+    it("returns 409 when user already has pending install in another org", async () => {
+      getPendingForgeInstallationForUserInOtherOrgMock.mockResolvedValueOnce({
+        id: "fgi_pending_other",
+        orgId: "org_2",
+      })
+
+      const app = createApp()
+      const res = await app.request("/connectors/atlassian/installation", {
+        method: "POST",
+      })
+
+      expect(res.status).toBe(409)
+      expect(await res.json()).toMatchObject({
+        code: "atlassian_pending_installation_exists",
+      })
+      expect(upsertPendingForgeInstallationMock).not.toHaveBeenCalled()
+    })
+
+    it("creates or updates pending install intent", async () => {
+      const app = createApp()
+      const res = await app.request("/connectors/atlassian/installation", {
+        method: "POST",
+      })
+
+      expect(res.status).toBe(200)
+      expect(upsertPendingForgeInstallationMock).toHaveBeenCalledWith({
+        orgId: "org_1",
+        installedByUserId: "user_1",
+      })
+    })
   })
 })
