+++ apps/backend/src/routes/v1/connectors-atlassian.ts
@@ -4,10 +4,12 @@ import { resolveAtlassianConfluenceApiBaseUrl } from "../../lib/atlassian-api-ba
 import {
   getAtlassianUserAccessToken,
   getForgeInstallationByOrgId,
+  getPendingForgeInstallationForUserInOtherOrg,
   listConfluenceSelectionsByOrgId,
   replaceConfluenceSelections,
+  type ForgeInstallation,
+  upsertPendingForgeInstallation,
 } from "../../models/atlassian-connector.js"
-import { atlassianInstallationRoutes } from "./atlassian-installation.js"
 
 const ErrorResponseSchema = z
   .object({
@@ -69,6 +71,42 @@ const SaveConfluenceSelectionResponseSchema = z
     savedCount: z.number(),
   })
   .openapi("SaveConfluenceSelectionResponse")
+
+const AtlassianInstallationSchema = z
+  .object({
+    id: z.string(),
+    orgId: z.string(),
+    cloudId: z.string().nullable(),
+    status: z.string(),
+    installedByUserId: z.string().nullable(),
+    createdAt: z.string().datetime(),
+    updatedAt: z.string().datetime(),
+  })
+  .openapi("AtlassianInstallation")
+
+const registerAtlassianInstallationRoute = createRoute({
+  method: "post",
+  path: "/installation",
+  responses: {
+    200: {
+      content: {
+        "application/json": {
+          schema: AtlassianInstallationSchema,
+        },
+      },
+      description: "Atlassian install intent registered for this org",
+    },
+    401: {
+      content: { "application/json": { schema: ErrorResponseSchema } },
+      description: "Unauthorized",
+    },
+    409: {
+      content: { "application/json": { schema: ErrorResponseSchema } },
+      description:
+        "Atlassian account is not linked or user already has pending install intent in another org",
+    },
+  },
+})
 
 const getStatusRoute = createRoute({
   method: "get",
@@ -210,7 +248,74 @@ async function fetchConfluence(
 }
 
 export const atlassianConnectorRoutes = new OpenAPIHono<AppEnv>()
-  .route("/installation", atlassianInstallationRoutes)
+  .openapi(registerAtlassianInstallationRoute, async (c) => {
+    if (!c.get("user") || !c.get("session")) {
+      return c.json({ error: "Unauthorized" }, 401)
+    }
+    const orgId = c.get("orgId")
+    if (!orgId) return c.json({ error: "Unauthorized" }, 401)
+
+    const user = c.get("user") as { id: string }
+    const accessToken = await getAtlassianUserAccessToken(user.id)
+    if (!accessToken) {
+      return c.json(
+        {
+          error: "Atlassian account not linked",
+          code: "atlassian_not_linked",
+        },
+        409,
+      )
+    }
+
+    const pendingInOtherOrg = await getPendingForgeInstallationForUserInOtherOrg({
+      userId: user.id,
+      orgId,
+    })
+    if (pendingInOtherOrg) {
+      return c.json(
+        {
+          error:
+            "A pending Atlassian installation already exists for this user in another organization",
+          code: "atlassian_pending_installation_exists",
+        },
+        409,
+      )
+    }
+
+    let row: ForgeInstallation
+    try {
+      row = await upsertPendingForgeInstallation({
+        orgId,
+        installedByUserId: user.id,
+      })
+    } catch (error) {
+      const dbError = error as { code?: string } | undefined
+      if (dbError?.code === "23505") {
+        return c.json(
+          {
+            error:
+              "A pending Atlassian installation already exists for this user in another organization",
+            code: "atlassian_pending_installation_exists",
+          },
+          409,
+        )
+      }
+      throw error
+    }
+
+    return c.json(
+      {
+        id: row.id,
+        orgId: row.orgId,
+        cloudId: row.cloudId ?? null,
+        status: row.status,
+        installedByUserId: row.installedByUserId ?? null,
+        createdAt: row.createdAt.toISOString(),
+        updatedAt: row.updatedAt.toISOString(),
+      },
+      200,
+    )
+  })
   .openapi(getStatusRoute, async (c) => {
     if (!c.get("user") || !c.get("session")) {
       return c.json({ error: "Unauthorized" }, 401)
